fix(product): added `Boxlink.js` for rendering md files
