+++ src/views/zesty/ProductLandingPage.js
@@ -23,13 +23,7 @@
  * Images API: https://zesty.org/services/media-storage-micro-dam/on-the-fly-media-optimization-and-dynamic-image-manipulation
  */
 import revampTheme from 'theme/revampTheme';
-import ModalImage from 'react-modal-image';
-
 import { v4 as uuidv4 } from 'uuid';
-
-import remarkGfm from 'remark-gfm';
-import ReactMarkdown from 'react-markdown';
-
 import { Box, Button, ThemeProvider } from '@mui/material';
 import {
   Link,
@@ -53,6 +47,7 @@ import { SearchModal } from 'views/Docs/SearchModal';
 import { AlgoSearch } from 'views/Docs/AlgoSearch';
 import { useZestyStore } from 'store';
 import GetDemoSection from 'revamp/ui/GetDemoSection';
+import { BoxLink } from 'components/markdown-styling/BoxLink';
 
 const GetTree = ({ data = [] }) => {
   return data.map((e) => {
@@ -266,7 +261,7 @@ const ProductLandingPage = (props) => {
   });
   const isLoggedIn = useIsLoggedIn();
   const { navData } = parseMarkdownFile({
-    markdown: content?.body || '',
+    markdown: content.body,
     tags: ['h2', 'h3', 'h4', 'h1', 'h5'],
     parentURL: '',
     title: '',
@@ -338,6 +333,13 @@ const ProductLandingPage = (props) => {
   };
   groupByUri(productsData);
 
+  // group the
+  const productGlossary = content.zesty.productGlossary.map((e) => {
+    const res = e.keywords.split(',').map((item) => item.toLowerCase());
+    return { ...e, target_words: res };
+  });
+  const mainKeywords = productGlossary.flatMap((obj) => obj.target_words);
+
   React.useEffect(() => {
     setalgoliaApiKey(props.content.algolia.apiKey);
     setalgoliaAppId(props.content.algolia.appId);
@@ -360,15 +362,38 @@ const ProductLandingPage = (props) => {
           justifyContent={'space-between'}
           alignItems={'center'}
           sx={{
-            display: { xs: 'none', md: 'flex' },
+            display: { xs: '', md: 'flex' },
           }}
         >
           <AppBar />
 
-          <SearchModal sx={{ width: true ? 300 : 500 }}>
+          <SearchModal
+            sx={{
+              width: true ? 300 : 500,
+              display: { xs: 'none', md: 'block' },
+            }}
+          >
             <AlgoSearch />
           </SearchModal>
         </Stack>
+
+        {/* Navigation mobile */}
+        <Stack
+          direction={'row'}
+          width={1}
+          justifyContent={'space-between'}
+          alignItems={'center'}
+          sx={{
+            display: { xs: '', md: 'none' },
+          }}
+        >
+          <Box>
+            <TreeNav
+              data={[{ title: 'Products', children: navigationData, uri: '#' }]}
+            />
+          </Box>
+        </Stack>
+
         {/* // body */}
         <Stack>
           <Grid container spacing={2} minHeight={'80vh'}>
@@ -397,111 +422,13 @@ const ProductLandingPage = (props) => {
                     {content?.title}
                   </Typography>
                 </Stack>
-                {/* <Stack width={1}>Tags Tags Tags</Stack> */}
                 <Stack width={1} height={1}>
-                  <ReactMarkdown
-                    remarkPlugins={[remarkGfm]}
-                    components={{
-                      img({ node }) {
-                        if (node.properties.title) {
-                          return (
-                            <Box px={5}>
-                              <ModalImage
-                                small={node.properties.src}
-                                large={node.properties.src}
-                                alt={node.properties.alt}
-                                title={node.properties.title}
-                              />
-                            </Box>
-                          );
-                        }
-                        return (
-                          <Box
-                            ml={3}
-                            mr={2}
-                            maxWidth={'200px'}
-                            sx={{
-                              float: 'right',
-                            }}
-                          >
-                            <ModalImage
-                              small={node.properties.src}
-                              large={node.properties.src}
-                              alt={node.properties.alt}
-                              title={node.properties.title}
-                            />
-                          </Box>
-                        );
-                      },
-                      blockquote({ node }) {
-                        return (
-                          <Box
-                            sx={{
-                              mx: 5,
-                              background: '#e7e7e7',
-                              p: 2,
-                              borderLeft: '2px #ccc solid',
-                              mb: 3,
-                            }}
-                          >
-                            <Typography>
-                              {node.children[1].children[0].value}
-                            </Typography>
-                          </Box>
-                        );
-                      },
-                      // h1({ node }) {
-                      //   return (
-                      //     <Box sx={{}}>
-                      //       <Typography
-                      //         variant="h4"
-                      //         component={'h1'}
-                      //         id={node.children[0].value
-                      //           ?.replace(/[^\w\s]/gi, '')
-                      //           ?.replace(/\s+/g, '-')
-                      //           ?.toLowerCase()}
-                      //       >
-                      //         {node.children[0].value}
-                      //       </Typography>
-                      //     </Box>
-                      //   );
-                      // },
-                      h2({ node }) {
-                        return (
-                          <Box sx={{}}>
-                            <Typography
-                              variant="h4"
-                              component={'h2'}
-                              id={node.children[0].value
-                                ?.replace(/[^\w\s]/gi, '')
-                                ?.replace(/\s+/g, '-')
-                                ?.toLowerCase()}
-                            >
-                              {node.children[0].value}
-                            </Typography>
-                          </Box>
-                        );
-                      },
-                      h3({ node }) {
-                        return (
-                          <Box sx={{}}>
-                            <Typography
-                              variant="h5"
-                              component={'h3'}
-                              id={node.children[0].value
-                                ?.replace(/[^\w\s]/gi, '')
-                                ?.replace(/\s+/g, '-')
-                                ?.toLowerCase()}
-                            >
-                              {node.children[0].value}
-                            </Typography>
-                          </Box>
-                        );
-                      },
-                    }}
-                  >
-                    {content.body}
-                  </ReactMarkdown>
+                  {/* Component that render the markdown file */}
+                  <BoxLink
+                    markdown={content.body}
+                    mainKeywords={mainKeywords}
+                    productGlossary={productGlossary}
+                  />
                 </Stack>
               </Stack>
             </Grid>
